pyproject.toml
[project]
name = "poetry-reader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.9"
dependencies = [
    "autocorrect>=2.6.1",
    "google-cloud-bigquery>=3.34.0",
    "google-cloud-storage>=3.1.1",
    "google-cloud-vision>=3.10.2",
    "pt-core-news-lg",
    "pyyaml>=6.0.2",
    "setuptools>=80.9.0",
    "spacy>=3.8.7",
    "whl>=0.0.4",
]

[tool.uv.sources]
pt-core-news-lg = { url = "https://github.com/explosion/spacy-models/releases/download/pt_core_news_lg-3.8.0/pt_core_news_lg-3.8.0-py3-none-any.whl" }

[dependency-groups]
dev = [
    "black>=25.1.0",
    "pytest>=8.4.1",
]

[tool.pytest.ini_options]
pythonpath = [
    'app', 'src'
    ]
